+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Changelog
 
 - [Changelog](#changelog)
+  - [0.1.1](#011)
   - [0.1.0](#010)
+
+## 0.1.1
+
+Released on 22/10/2024
+
+- `with_get_uid` constructor
+- `with_get_gid` constructor
 
 ## 0.1.0